Split p2p classifier and route-cap tests
