+++ crates/oasis7_net/src/libp2p_net.rs
@@ -111,9 +111,7 @@ const RR_GET_LOCAL_PEER_RECORD: &str = "/aw/rr/1.0.0/get_local_peer_record";
 const RR_GET_CACHED_PEER_RECORD: &str = "/aw/rr/1.0.0/get_cached_peer_record";
 const RR_GET_CACHED_DISCOVERY_PEERS: &str = "/aw/rr/1.0.0/get_cached_discovery_peers";
 const LIFECYCLE_EVENT_ERROR_COOLDOWN_MS: i64 = 5_000;
-
 type CommandResponseSender<T> = std::sync::mpsc::Sender<Result<T, WorldError>>;
-
 #[derive(Clone)]
 pub struct Libp2pNetwork {
     peer_id: PeerId,
@@ -130,9 +128,7 @@ pub struct Libp2pNetwork {
     traffic_metrics: SharedLibp2pTrafficMetrics,
     wire_byte_counters: SharedLibp2pWireByteCounters,
 }
-
 type Handler = Arc<dyn Fn(&[u8]) -> Result<Vec<u8>, WorldError> + Send + Sync>;
-
 impl Libp2pNetwork {
     pub fn new(config: Libp2pNetworkConfig) -> Self {
         let keypair = config
@@ -157,7 +153,6 @@ impl Libp2pNetwork {
         let max_published_messages = config.max_published_messages.max(1);
         let max_error_messages = config.max_error_messages.max(1);
         let max_listening_addrs = config.max_listening_addrs.max(1);
-
         let event_inbox = Arc::clone(&inbox);
         let event_published = Arc::clone(&published);
         let event_listening_addrs = Arc::clone(&listening_addrs);
@@ -182,7 +177,6 @@ impl Libp2pNetwork {
                 .as_ref()
                 .map(peer_record_enables_rendezvous)
                 .unwrap_or(false);
-
         std::thread::spawn(move || {
             let mut swarm = build_swarm(
                 &keypair_clone,
@@ -246,7 +240,6 @@ impl Libp2pNetwork {
                     push_bounded_clone(&event_errors, msg, max_error_messages, "lock errors");
                 }
             }
-
             schedule_periodic_republish(republish_tx, republish_interval_ms);
             if peer_record_template.is_some() {
                 schedule_periodic_discovery_refresh(discovery_tx, discovery_query_interval_ms);
@@ -256,7 +249,6 @@ impl Libp2pNetwork {
                 bootstrap_redial_peers,
                 bootstrap_redial_interval_ms,
             );
-
             async_std::task::block_on(async move {
                 let mut command_rx = command_rx;
                 let command_ctx = CommandContext {
@@ -1197,7 +1189,6 @@ impl Drop for Libp2pNetwork {
         let _ = self.enqueue_command(Command::Shutdown);
     }
 }
-
 #[cfg(test)]
 mod tests;
 #[cfg(test)]
