+++ crates/oasis7_node/src/network_bridge.rs
@@ -5,7 +5,7 @@ use std::sync::{Arc, Mutex};
 use std::time::{Duration, Instant};
 
 use oasis7_net::{world_error_is_publish_failure, world_error_is_retryable_connection_gap};
-use oasis7_proto::distributed::{DistributedErrorCode, WorldHeadAnnounce};
+use oasis7_proto::distributed::WorldHeadAnnounce;
 use oasis7_proto::distributed_dht as proto_dht;
 use oasis7_proto::distributed_net::{
     classify_network_protocol, DistributedNetwork, NetworkLane, NetworkLaneOperation,
@@ -17,6 +17,13 @@ use serde::Serialize;
 
 use crate::gossip_udp::{
     GossipAttestationMessage, GossipCommitMessage, GossipMessage, GossipProposalMessage,
+};
+pub(crate) use crate::network_error_classification::{
+    replication_network_error_is_availability_gap, replication_network_error_is_not_found,
+    replication_network_error_is_protocol_unavailable,
+    replication_network_error_is_route_unavailable, replication_network_error_is_timeout_protocol,
+    replication_network_error_is_unsupported_protocol, replication_network_error_kind_label,
+    replication_network_error_mentions_protocol,
 };
 use crate::replication::{
     load_blob_from_root, FetchCommitRequest, FetchCommitResponse, FetchHeadRequest,
@@ -1148,90 +1155,4 @@ fn replication_network_error_detail(err: &WorldError) -> &str {
         WorldError::Io(message) | WorldError::Serde(message) => message.as_str(),
         WorldError::SignatureKeyInvalid => "invalid signature key",
     }
-}
-
-fn replication_network_error_kind_label(code: DistributedErrorCode) -> &'static str {
-    match code {
-        DistributedErrorCode::ErrNotFound => "not_found",
-        DistributedErrorCode::ErrUnsupported => "unsupported",
-        DistributedErrorCode::ErrTimeout => "timeout",
-        DistributedErrorCode::ErrNotAvailable => "not_available",
-        DistributedErrorCode::ErrBusy => "busy",
-        DistributedErrorCode::ErrRateLimited => "rate_limited",
-        DistributedErrorCode::ErrBadRequest => "bad_request",
-        DistributedErrorCode::ErrUnauthorized => "unauthorized",
-        DistributedErrorCode::ErrStateMismatch => "state_mismatch",
-        DistributedErrorCode::ErrInvalidHash => "invalid_hash",
-    }
-}
-
-pub(crate) fn replication_network_error_is_availability_gap(err: &NodeError) -> bool {
-    let NodeError::Replication { reason } = err else {
-        return false;
-    };
-    reason.starts_with(REPLICATION_NETWORK_AVAILABILITY_GAP_PREFIX)
-}
-
-pub(crate) fn replication_network_error_is_route_unavailable(err: &NodeError) -> bool {
-    let NodeError::Replication { reason } = err else {
-        return false;
-    };
-    reason.starts_with(REPLICATION_NETWORK_ROUTE_UNAVAILABLE_PREFIX)
-}
-
-pub(crate) fn replication_network_error_mentions_protocol(err: &NodeError, protocol: &str) -> bool {
-    let NodeError::Replication { reason } = err else {
-        return false;
-    };
-    reason.contains(protocol)
-}
-
-pub(crate) fn replication_network_error_is_not_found(err: &NodeError) -> bool {
-    let NodeError::Replication { reason } = err else {
-        return false;
-    };
-    network_request_reason_has_kind(reason, "not_found")
-        || (reason.contains("NetworkRequestFailed") && reason.contains("ErrNotFound"))
-}
-
-pub(crate) fn replication_network_error_is_unsupported_protocol(
-    err: &NodeError,
-    protocol: &str,
-) -> bool {
-    let NodeError::Replication { reason } = err else {
-        return false;
-    };
-    reason.contains(protocol)
-        && (network_request_reason_has_kind(reason, "unsupported")
-            || (reason.contains("NetworkRequestFailed") && reason.contains("ErrUnsupported")))
-}
-
-pub(crate) fn replication_network_error_is_protocol_unavailable(
-    err: &NodeError,
-    protocol: &str,
-) -> bool {
-    let NodeError::Replication { reason } = err else {
-        return false;
-    };
-    (reason.starts_with(REPLICATION_NETWORK_ROUTE_UNAVAILABLE_PREFIX) && reason.contains(protocol))
-        || reason.contains("NetworkProtocolUnavailable")
-            && (reason.contains("handler missing") || reason.contains(protocol))
-}
-
-pub(crate) fn replication_network_error_is_timeout_protocol(
-    err: &NodeError,
-    protocol: &str,
-) -> bool {
-    let NodeError::Replication { reason } = err else {
-        return false;
-    };
-    reason.contains(protocol)
-        && (network_request_reason_has_kind(reason, "timeout")
-            || reason.contains("request failed: Timeout"))
-}
-
-fn network_request_reason_has_kind(reason: &str, kind: &str) -> bool {
-    reason
-        .split_whitespace()
-        .any(|field| field.strip_prefix("kind=") == Some(kind))
 }
